Refactoring WIP
