+++ lib/core/src/display.rs
@@ -1,6 +1,8 @@
 pub mod camera;
 pub mod rendering;
 pub mod scene;
+pub mod shape;
 pub mod symbol;
+pub mod object;
 pub mod workspace;
 pub mod world;
