+++ lib/core/src/display/camera.rs
@@ -2,7 +2,7 @@
 use crate::prelude::*;
 
 use crate::dirty;
-use crate::display::symbol::DisplayObjectData;
+use crate::display::object::DisplayObjectData;
 use nalgebra::{Vector3, Vector4, Matrix4, Perspective3};
 use basegl_system_web::Logger;
 use crate::dirty::traits::*;
