+++ lib/core/src/display/symbol/geometry.rs
