+++ lib/core/src/display/symbol/geometry/primitive.rs
