+++ lib/core/src/display/symbol/geometry/primitive/mesh.rs
