+++ lib/core/src/display/symbol/geometry/primitive/mesh/buffer/item.rs
