+++ lib/core/src/display/symbol/geometry/primitive/mesh/scope.rs
@@ -6,9 +6,9 @@ use crate::data::function::callback::*;
 use crate::data::opt_vec::OptVec;
 use crate::dirty;
 use crate::dirty::traits::*;
-use crate::object::geometry::primitive::mesh::buffer;
-use crate::object::geometry::primitive::mesh::buffer::item::Item;
-use crate::object::geometry::primitive::mesh::buffer::IsBuffer;
+use crate::display::symbol::geometry::primitive::mesh::buffer;
+use crate::display::symbol::geometry::primitive::mesh::buffer::item::Item;
+use crate::display::symbol::geometry::primitive::mesh::buffer::IsBuffer;
 use crate::system::web::group;
 use crate::system::web::Logger;
 use crate::promote;
