+++ lib/core/src/display/symbol/material.rs
