+++ lib/core/src/display/symbol/material/shader.rs
