+++ lib/core/src/display/symbol/material/shader/builder.rs
@@ -2,7 +2,7 @@ use crate::prelude::*;
 
 use super::glsl;
 use crate::data::container::Add;
-use crate::object::material::shader::glsl::{Type, PrimType};
+use crate::display::symbol::material::shader::glsl::{Type, PrimType};
 use code_builder::HasCodeRepr;
 use std::collections::BTreeMap;
