+++ lib/core/src/display/symbol/material/shader/glsl.rs
